pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-harness"
version = "1.0.0"
description = "An agent orchestration harness for multi-model AI workflows"
requires-python = ">=3.11"
dependencies = [
    "langgraph",
    "litellm",
    "python-telegram-bot",
    "arize-phoenix",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "kuzu",
    "pyyaml",
    "pydantic",
    "httpx",
    "aiohttp",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
harness-ctl = "apps.orchestrator.cli:main"

[project.entry-points."apps.orchestrator"]
main = "apps.orchestrator.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
